+++ src/spyglass/spikesorting/v1/__init__.py
@@ -19,4 +19,3 @@ from .recording import (
     SpikeSortingRecordingSelection,
 )
 from .sorting import SpikeSorterParameters, SpikeSorting, SpikeSortingSelection
-from .utils import get_spiking_sorting_v1_merge_ids